i mean i sent you a message on whatsapp at 00:30 on whatsapp so yeah check that but ftr: Downloader reads the directory, and added arguments to function Downloader and files are only downaded once when they have the same artist and song name to save time

+++ Downloader.py
@@ -5,33 +5,37 @@ import urllib.parse
 import re
 import eyed3
 import os
+import os.path
 import subprocess 
 from subprocess import call
-default = 'default'
+default = 'default'#literally is a placeholder not sure if the variable does anything
 check = []
 q = 0
 
-def download(artist, title, track_num, album_artist, album, playlistname):
+def download(artist, title, track_num, album_artist, album, playlistname, number_items):
     filename = setfilename(title,artist)
-    hold = album
+    global q
     #cwd = os.getcwd() + "/output/" #os.getcwd() returns the current diretory
-    output_folder = os.getcwd() + "\\" + 'output' + "\\"
-    query_string = urllib.parse.urlencode({"search_query" : artist + " " + title})  #where the user input
-    html_content = urllib.request.urlopen("http://www.youtube.com/results?" + query_string)
-    search_results = re.findall(r'href=\"\/watch\?v=(.{11})', html_content.read().decode())
-    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-        ydl.download(['http://www.youtube.com/watch?v=' + search_results[0]])
-    audiofile = eyed3.load(output_folder+"NA.mp3")
-    audiofile.tag.artist = artist
-    audiofile.tag.album = album
-    audiofile.tag.album_artist = album_artist
-    audiofile.tag.title = title
-    audiofile.tag.track_num = track_num
-    audiofile.tag.save()
-    os.rename(output_folder+"NA.mp3", output_folder+filename)
+    output_folder = os.getcwd() + "\\" + 'NA' + "\\"  #should i make this global rather than have it done every time the function is called?
+    if not(os.path.isfile(output_folder+filename)) and not(os.path.isfile(os.getcwd()+ "\\"+playlistname+"\\"+filename)):
+        query_string = urllib.parse.urlencode({"search_query" : artist + " " + title})  #where the user input
+        html_content = urllib.request.urlopen("http://www.youtube.com/results?" + query_string)
+        search_results = re.findall(r'href=\"\/watch\?v=(.{11})', html_content.read().decode())
+        with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+            ydl.download(['http://www.youtube.com/watch?v=' + search_results[0]])
+        audiofile = eyed3.load(output_folder+"NA.mp3")
+        audiofile.tag.artist = artist
+        audiofile.tag.album = album
+        audiofile.tag.album_artist = album_artist
+        audiofile.tag.title = title
+        audiofile.tag.track_num = track_num
+        audiofile.tag.save()
+        os.rename(output_folder+"NA.mp3", output_folder+filename)
+    if q == number_items:
+        renameoutputfolder(playlistname)	
 
 ydl_opts = {
-    'outtmpl' : '/' + 'output' + '/%(default)s.%(ext)s', 'format': 'bestaudio/best',
+    'outtmpl' : '/' + '%(default)s' + '/%(default)s.%(ext)s', 'format': 'bestaudio/best',
     'postprocessors': [{
         'key': 'FFmpegExtractAudio',
         'preferredcodec': 'mp3',
@@ -41,14 +45,13 @@ ydl_opts = {
 
 def setfilename(back, lead):
     global check, q
-    collect = lead + " - " + back
-    if q>0:
-        for x in range (q):#reads 0 upto q
-            if check[x] == collect:
-                print("%d" % x)
-                print("Repeated file: %s\n" % collect)
-                collect += "(1)"
+    collect = lead + " - " + back         
     check.append(collect)
     q+=1
     filename = collect +'.mp3'
     return filename
+ 	
+
+def renameoutputfolder(playlistname):
+	output_folder = os.getcwd()
+	os.rename(output_folder+"/NA", output_folder+"/"+playlistname)
